package.json: Enforce c++14 for native module builds
Via various dependencies, we require a higher version of node now, which uses a
version of V8 that requires c++14. For this reason, most native modules will
also need to be built with c++14 set as the baseline. While most of our native
modules do so, usb does not, so we need to work that around.

To this end, add `CXXFLAGS=-std=c++14` to our `postinstall` script, which makes
usb build too.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>